Allow adding&deleting granules from subdirs
If we're defining multiple layers with the same config file,
it might make sense to have the files in subdirectories per layer
this allows keeping the subdir specified in argument
or keeping the path from the geoserver granule location

Note! Deleting files with this might fail if the path
in geoserver is different than path on host server!